+++ gui_leptos_tauri/src-tauri/src/prim_mst.rs
@@ -7,7 +7,7 @@ use crate::grid::{Net, Point, PointPair};
 
 
 
-pub fn prim_mst(edges: Vec<(OrderedFloat<f32>, Net, PointPair)>) -> Vec<(OrderedFloat<f32>, Net, PointPair)> {
+pub fn prim_mst(edges: Vec<(OrderedFloat<f64>, Net, PointPair)>) -> Vec<(OrderedFloat<f64>, Net, PointPair)> {
     assert!(edges.len() >=1, "At least 1 edge is required to form a minimum spanning tree.");
     let (_, net, _) = edges[0].clone();
     let mut index_to_point: BTreeSet<Point> = BTreeSet::new();
@@ -21,14 +21,14 @@ pub fn prim_mst(edges: Vec<(OrderedFloat<f32>, Net, PointPair)>) -> Vec<(Ordered
         .enumerate()
         .map(|(i, point)| (*point, i as i32))
         .collect();
-    let edges: Vec<(OrderedFloat<f32>, Net, i32, i32)> = edges.into_iter()
+    let edges: Vec<(OrderedFloat<f64>, Net, i32, i32)> = edges.into_iter()
         .map(|(weight, net, point_pair)| {
             let start_index = point_to_index[&point_pair.start()];
             let end_index = point_to_index[&point_pair.end()];
             (weight, net, start_index, end_index)
         })
         .collect();
-    let undirected_graph: Vec<Vec<(OrderedFloat<f32>, i32)>> = {
+    let undirected_graph: Vec<Vec<(OrderedFloat<f64>, i32)>> = {
         let mut graph = vec![vec![]; point_to_index.len()];
         for (weight, net, start, end) in &edges {
             graph[*start as usize].push((weight.clone(), *end));
@@ -36,12 +36,12 @@ pub fn prim_mst(edges: Vec<(OrderedFloat<f32>, Net, PointPair)>) -> Vec<(Ordered
         }
         graph
     };
-    let mut mst_edges: Vec<(OrderedFloat<f32>, Net, i32, i32)> = vec![];
+    let mut mst_edges: Vec<(OrderedFloat<f64>, Net, i32, i32)> = vec![];
     let mut visited = std::collections::HashSet::new();
     // let edges = edges.into_iter()
     //     .map(|(weight, net, start, end)| (weight, Option<(Net, Point, Point)>::Some((net, start, end))))
     //     .collect::<Vec<_>>();
-    let mut priority_queue: BinaryHeap<Reverse<(OrderedFloat<f32>, i32, i32)>> = BinaryHeap::new();
+    let mut priority_queue: BinaryHeap<Reverse<(OrderedFloat<f64>, i32, i32)>> = BinaryHeap::new();
     priority_queue.push(Reverse((OrderedFloat(0.0), 0, -1))); // Start with a dummy edge
     while let Some(Reverse((weight, u, v))) = priority_queue.pop() {
         if visited.contains(&u) {
@@ -67,7 +67,7 @@ pub fn prim_mst(edges: Vec<(OrderedFloat<f32>, Net, PointPair)>) -> Vec<(Ordered
     // let mst_edges = mst_edges.into_iter()
     //     .map(|((start, end), (weight, net))| (weight, net, start, end))
     //     .collect::<Vec<_>>();
-    let mst_edges: Vec<(OrderedFloat<f32>, Net, PointPair)> = mst_edges.into_iter()
+    let mst_edges: Vec<(OrderedFloat<f64>, Net, PointPair)> = mst_edges.into_iter()
         .map(|(weight, net, start, end)| {
             let start_point = index_to_point[start as usize].clone();
             let end_point = index_to_point[end as usize].clone();
